# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "Movies/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update movies",
                "description": "-",
                "operationId": "movies/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "overview",
                                    "run_time",
                                    "release_date",
                                    "total_likes",
                                    "total_ratings",
                                    "average_rating",
                                    "poster_image",
                                    "cover_image",
                                    "trailer_url",
                                    "last_upload_date",
                                    "subscription_only",
                                    "expired_subscription_only"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "overview": {
                                        "type": "string"
                                    },
                                    "run_time": {
                                        "type": "integer"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_likes": {
                                        "type": "integer"
                                    },
                                    "total_ratings": {
                                        "type": "integer"
                                    },
                                    "average_rating": {
                                        "type": "string"
                                    },
                                    "poster_image": {
                                        "type": "string"
                                    },
                                    "cover_image": {
                                        "type": "string"
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    },
                                    "last_upload_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "subscription_only": {
                                        "type": "string"
                                    },
                                    "expired_subscription_only": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete movies",
                "description": "-",
                "operationId": "movies/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/requested-movies": {
            "get": {
                "tags": [
                    "Requested-Movies"
                ],
                "summary": "Get List requested-movies Data",
                "description": "enter your requested-movies here",
                "operationId": "requested-movies",
                "responses": {
                    "default": {
                        "description": "return array model requested-movies"
                    }
                }
            },
            "post": {
                "tags": [
                    "Requested-Movies"
                ],
                "summary": "requested-movies",
                "description": "requested-movies",
                "operationId": "Requested-Movies",
                "requestBody": {
                    "description": "form requested-movies",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "description",
                                    "image_path",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "artist_profile": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/requested-movies/{id}": {
            "get": {
                "tags": [
                    "Requested-Movies"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "requested-movies/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Requested-Movies"
                ],
                "summary": "Update requested-movies",
                "description": "-",
                "operationId": "requested-movies/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "description",
                                    "image_path",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "artist_profile": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Requested-Movies"
                ],
                "summary": "Delete requested-movies",
                "description": "-",
                "operationId": "requested-movies/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}